requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
google-cloud-storage
pyyaml